move routerFactory to routes
