+++ src/routes/authentication/index.ts
@@ -1,6 +1,6 @@
 import mU from "@src/messages/constants/user";
 import mC from "@src/messages/constants/otp";
-import { postRouteFactory } from "@src/utilities/routes_factory/post";
+import { postRouteFactory } from "@src/routes/_routesFactory/post";
 import { authenticateUser } from "@src/services/controllers/authentication";
 
 export const postAuthUser = postRouteFactory(
