+++ src/routes/sendemail/index.ts
@@ -1,7 +1,7 @@
 import mU from "@src/messages/constants/user";
 import mE from "@src/messages/constants/email";
 import { sendOTPEmail } from "@src/services/controllers/sendemail";
-import { postRouteFactory } from "@src/utilities/routes_factory/post";
+import { postRouteFactory } from "@src/routes/_routesFactory/post";
 
 export const postSendOTPEmail = postRouteFactory(
   mE.email,
